# docs/openapi-v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "乒乓球赛事编排与赛务管理系统 Demo",
    "version": "0.1.0"
  },
  "paths": {
    "/api/tournaments": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "summary": "List Tournaments",
        "operationId": "list_tournaments_api_tournaments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TournamentOut"
                  },
                  "type": "array",
                  "title": "Response List Tournaments Api Tournaments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tournaments"
        ],
        "summary": "Create Tournament",
        "operationId": "create_tournament_api_tournaments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}": {
      "get": {
        "tags": [
          "tournaments"
        ],
        "summary": "Get Tournament",
        "operationId": "get_tournament_api_tournaments__tournament_id__get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tournaments"
        ],
        "summary": "Delete Tournament",
        "description": "删除赛事（级联删除分组/选手/球台/比赛，见 db.py 外键 ON DELETE CASCADE）。",
        "operationId": "delete_tournament_api_tournaments__tournament_id__delete",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/matches/{match_id}/assign-table": {
      "post": {
        "tags": [
          "scheduling"
        ],
        "summary": "Assign Table",
        "operationId": "assign_table_api_matches__match_id__assign_table_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/release": {
      "post": {
        "tags": [
          "scheduling"
        ],
        "summary": "Release Match",
        "operationId": "release_match_api_matches__match_id__release_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/schedule-next": {
      "post": {
        "tags": [
          "scheduling"
        ],
        "summary": "Schedule Next",
        "operationId": "schedule_next_api_tournaments__tournament_id__schedule_next_post",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleNextResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/dashboard": {
      "get": {
        "tags": [
          "scheduling"
        ],
        "summary": "Dashboard",
        "operationId": "dashboard_api_tournaments__tournament_id__dashboard_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dashboard"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/score": {
      "post": {
        "tags": [
          "scores"
        ],
        "summary": "Record Score",
        "operationId": "record_score_api_matches__match_id__score_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches/{match_id}/revise-score": {
      "post": {
        "tags": [
          "scores"
        ],
        "summary": "Revise Score",
        "operationId": "revise_score_api_matches__match_id__revise_score_post",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tournaments/{tournament_id}/rankings": {
      "get": {
        "tags": [
          "scores"
        ],
        "summary": "Get Rankings",
        "operationId": "get_rankings_api_tournaments__tournament_id__rankings_get",
        "parameters": [
          {
            "name": "tournament_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Tournament Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RankingsResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AssignTableRequest": {
        "properties": {
          "table_id": {
            "type": "integer",
            "title": "Table Id"
          }
        },
        "type": "object",
        "required": [
          "table_id"
        ],
        "title": "AssignTableRequest"
      },
...
      "BronzeMode": {
        "type": "string",
        "enum": [
          "BRONZE_MATCH",
          "JOINT_BRONZE"
        ],
        "title": "BronzeMode"
      },
...
      "Dashboard": {
        "properties": {
          "tournament": {
            "$ref": "#/components/schemas/TournamentOut"
          },
          "stats": {
            "$ref": "#/components/schemas/DashboardStats"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableWithMatch"
            },
            "type": "array",
            "title": "Tables"
          },
          "next_playable": {
            "items": {
              "$ref": "#/components/schemas/MatchOut"
            },
            "type": "array",
            "title": "Next Playable"
          }
        },
        "type": "object",
        "required": [
          "tournament",
          "stats",
          "tables",
          "next_playable"
        ],
        "title": "Dashboard"
      },
      "DashboardStats": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "finished": {
            "type": "integer",
            "title": "Finished"
          },
          "playing": {
            "type": "integer",
            "title": "Playing"
          },
          "waiting": {
            "type": "integer",
            "title": "Waiting"
          }
        },
        "type": "object",
        "required": [
          "total",
          "finished",
          "playing",
          "waiting"
        ],
        "title": "DashboardStats"
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "SINGLES",
          "DOUBLES"
        ],
        "title": "EventType",
        "description": "参赛项目。第一版完整支持单打与固定搭档双打。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchBracket": {
        "type": "string",
        "enum": [
          "GROUP",
          "MAIN",
          "PLACEMENT"
        ],
        "title": "MatchBracket"
      },
...
      "MatchOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "tournament_id": {
            "type": "integer",
            "title": "Tournament Id"
          },
          "stage": {
            "$ref": "#/components/schemas/MatchStage"
          },
          "group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Id"
          },
          "round": {
            "type": "integer",
            "title": "Round"
          },
          "match_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Index"
          },
          "player_a_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player A Id"
          },
          "player_b_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player B Id"
          },
          "player_a_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player A Score"
          },
          "player_b_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player B Score"
          },
          "winner_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Id"
          },
          "table_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Table Id"
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "prev_match_a_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Match A Id"
          },
          "prev_match_b_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Match B Id"
          },
          "entry_a_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry A Id"
          },
          "entry_b_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry B Id"
          },
          "winner_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Entry Id"
          },
          "entry_a_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry A Name"
          },
          "entry_b_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry B Name"
          },
          "result_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultType"
              },
              {
                "type": "null"
              }
            ]
          },
          "forfeit_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forfeit Entry Id"
          },
          "bracket": {
            "$ref": "#/components/schemas/MatchBracket",
            "default": "GROUP"
          },
          "placement_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement Min"
          },
          "placement_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placement Max"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/MatchGameOut"
            },
            "type": "array",
            "title": "Games",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "tournament_id",
          "stage",
          "group_id",
          "round",
          "match_index",
          "player_a_id",
          "player_b_id",
          "player_a_score",
          "player_b_score",
          "winner_id",
          "table_id",
          "status",
          "prev_match_a_id",
          "prev_match_b_id"
        ],
        "title": "MatchOut"
      },
      "MatchStage": {
        "type": "string",
        "enum": [
          "GROUP",
          "KNOCKOUT"
        ],
        "title": "MatchStage",
        "description": "比赛所属赛段。"
      },
      "MatchStatus": {
        "type": "string",
        "enum": [
          "WAITING",
          "PLAYING",
          "FINISHED"
        ],
        "title": "MatchStatus",
        "description": "比赛状态。"
      },
...
      "PlacementMode": {
        "type": "string",
        "enum": [
          "OFF",
          "COMPLETE",
          "TIERED"
        ],
        "title": "PlacementMode"
      },
...
      "RankingsResult": {
        "properties": {
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/GroupRankingOut"
            },
            "type": "array",
            "title": "Rankings"
          }
        },
        "type": "object",
        "required": [
          "rankings"
        ],
        "title": "RankingsResult"
      },
      "ResultType": {
        "type": "string",
        "enum": [
          "NORMAL",
          "FORFEIT",
          "WALKOVER",
          "NO_SHOW",
          "DISQUALIFIED"
        ],
        "title": "ResultType"
      },
      "ScheduleNextResult": {
        "properties": {
          "assigned": {
            "type": "integer",
            "title": "Assigned"
          },
          "assignments": {
            "items": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Assignments"
          }
        },
        "type": "object",
        "required": [
          "assigned",
          "assignments"
        ],
        "title": "ScheduleNextResult"
      },
      "ScoreRequest": {
        "properties": {
          "player_a_score": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Player A Score"
          },
          "player_b_score": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Player B Score"
          },
          "games": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MatchGameInput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Games"
          },
          "result_type": {
            "$ref": "#/components/schemas/ResultType",
            "default": "NORMAL"
          },
          "forfeit_entry_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forfeit Entry Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "ScoreRequest"
      },
...
      "TournamentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "table_count": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Table Count"
          },
          "group_count": {
            "type": "integer",
            "maximum": 26.0,
            "minimum": 1.0,
            "title": "Group Count"
          },
          "qualify_per_group": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Qualify Per Group"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType",
            "default": "SINGLES"
          },
          "bronze_mode": {
            "$ref": "#/components/schemas/BronzeMode",
            "default": "JOINT_BRONZE"
          },
          "placement_mode": {
            "$ref": "#/components/schemas/PlacementMode",
            "default": "OFF"
          },
          "games_to_win": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Games To Win",
            "default": 2
          },
          "points_to_win": {
            "type": "integer",
            "maximum": 99.0,
            "minimum": 1.0,
            "title": "Points To Win",
            "default": 11
          }
        },
        "type": "object",
        "required": [
          "name",
          "date",
          "table_count",
          "group_count",
          "qualify_per_group"
        ],
        "title": "TournamentCreate"
      },
      "TournamentOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "table_count": {
            "type": "integer",
            "title": "Table Count"
          },
          "group_count": {
            "type": "integer",
            "title": "Group Count"
          },
          "qualify_per_group": {
            "type": "integer",
            "title": "Qualify Per Group"
          },
          "stage": {
            "$ref": "#/components/schemas/TournamentStage"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType",
            "default": "SINGLES"
          },
          "bronze_mode": {
            "$ref": "#/components/schemas/BronzeMode",
            "default": "JOINT_BRONZE"
          },
          "placement_mode": {
            "$ref": "#/components/schemas/PlacementMode",
            "default": "OFF"
          },
          "games_to_win": {
            "type": "integer",
            "title": "Games To Win",
            "default": 2
          },
          "points_to_win": {
            "type": "integer",
            "title": "Points To Win",
            "default": 11
          },
          "roster_confirmed": {
            "type": "boolean",
            "title": "Roster Confirmed",
            "default": false
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "date",
          "table_count",
          "group_count",
          "qualify_per_group",
          "stage",
          "created_at"
        ],
        "title": "TournamentOut"
      },
      "TournamentStage": {
        "type": "string",
        "enum": [
          "REGISTRATION",
          "GROUP_STAGE",
          "KNOCKOUT",
          "FINISHED"
        ],
        "title": "TournamentStage",
        "description": "赛事阶段。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}